Moved the storage service to a new file.
