+++ package/cloudshell/cp/azure/domain/services/storage_service.py
@@ -1,0 +1,18 @@
+import azure
+from azure.mgmt.storage.models import SkuName, StorageAccountCreateParameters
+
+
+class StorageService(object):
+    def __init__(self, storage_client):
+        self.storage_client = storage_client
+
+    def create_storage_account(self, group_name, region, storage_account_name):
+        kind_storage_value = azure.mgmt.storage.models.Kind.storage.value
+        sku_name = SkuName.standard_lrs
+        sku = azure.mgmt.storage.models.Sku(sku_name)
+        storage_accounts_create = self.storage_client.storage_accounts.create(group_name, storage_account_name,
+                                                                              StorageAccountCreateParameters(
+                                                                                  sku=sku,
+                                                                                  kind=kind_storage_value,
+                                                                                  location=region))
+        storage_accounts_create.wait()  # async operation
