+++ package/cloudshell/cp/azure/domain/services/virtual_machine_service.py
@@ -2,7 +2,7 @@ import azure
 from azure.mgmt.compute.models import OSProfile, HardwareProfile, VirtualMachineSizeTypes, NetworkProfile, \
     NetworkInterfaceReference, CachingTypes, DiskCreateOptionTypes, VirtualHardDisk, ImageReference, OSDisk
 from azure.mgmt.resource.resources.models import ResourceGroup
-from azure.mgmt.storage.models import StorageAccountCreateParameters, SkuName
+from azure.mgmt.storage.models import SkuName
 
 
 class VirtualMachineService(object):
@@ -72,17 +72,3 @@ class VirtualMachineService(object):
         )
 
 
-class StorageService(object):
-    def __init__(self, storage_client):
-        self.storage_client = storage_client
-
-    def create_storage_account(self, group_name, region, storage_account_name):
-        kind_storage_value = azure.mgmt.storage.models.Kind.storage.value
-        sku_name = SkuName.standard_lrs
-        sku = azure.mgmt.storage.models.Sku(sku_name)
-        storage_accounts_create = self.storage_client.storage_accounts.create(group_name, storage_account_name,
-                                                                              StorageAccountCreateParameters(
-                                                                                  sku=sku,
-                                                                                  kind=kind_storage_value,
-                                                                                  location=region))
-        storage_accounts_create.wait()  # async operation
